// lmss-fullstack/ClientApp/src/app/pages/users/model/user-response.model.ts
export interface UsersResponse {
  emailVerified: boolean;
  disabled: boolean;
  roles: [];
  _id: string;
  email: string;
  password: string;
  firstName: string;
  avatars: [];
  createdAt: Date;
  updatedAt: Date;
  __v: number;
  authenticationUid: string;
  lastName: string;
  phoneNumber: number;
  updatedBy: string;
  id: string;
}